+++ mex/backend/main.py
@@ -21,13 +21,13 @@ from mex.backend.logging import UVICORN_LOGGING_CONFIG
 from mex.backend.match.main import router as match_router
 from mex.backend.merged.main import router as merged_router
 from mex.backend.oauth import router as oauth_router
+from mex.backend.person.main import router as person_router
 from mex.backend.preview.main import router as preview_router
 from mex.backend.publishable_merged.main import router as publishable_merged_router
 from mex.backend.rules.main import router as rules_router
 from mex.backend.security import SWAGGER_UI_INIT_OAUTH
 from mex.backend.settings import BackendSettings
 from mex.backend.system.main import router as system_router
-from mex.backend.user.main import router as user_router
 from mex.common.cli import entrypoint
 from mex.common.connector import CONNECTOR_STORE
 from mex.common.logging import logger
@@ -86,7 +86,7 @@ router.include_router(preview_router)
 router.include_router(publishable_merged_router)
 router.include_router(rules_router)
 router.include_router(wikidata_router)
-router.include_router(user_router)
+router.include_router(person_router)
 router.include_router(oauth_router)
 router.include_router(system_router)
