+++ tests/person/test_main.py
@@ -8,7 +8,7 @@ if TYPE_CHECKING:  # pragma: no cover
     from fastapi.testclient import TestClient
 
 
-def test_get_current_user(
+def test_get_current_person(
     client_with_bearer_write_permission: TestClient,
 ) -> None:
     with (
@@ -17,9 +17,9 @@ def test_get_current_user(
             "mex.backend.security.jwt.decode",
             return_value={"preferred_username": "Writer", "groups": ["Abteilung_21"]},
         ),
-        patch("mex.backend.user.main.LDAPConnector.get") as mock_ldap_connector_get,
-        patch("mex.backend.user.main.get_provider") as mock_get_provider,
-        patch("mex.backend.user.main.get_merged_item") as mock_get_merged_item,
+        patch("mex.backend.person.main.LDAPConnector.get") as mock_ldap_connector_get,
+        patch("mex.backend.person.main.get_provider") as mock_get_provider,
+        patch("mex.backend.person.main.get_merged_item") as mock_get_merged_item,
     ):
         mock_jwks_client = MagicMock()
         mock_jwks_client.get_signing_key_from_jwt.return_value = MagicMock(
@@ -44,7 +44,7 @@ def test_get_current_user(
         )
         mock_get_merged_item.return_value = mock_person
 
-        response = client_with_bearer_write_permission.get("/v0/user/me")
+        response = client_with_bearer_write_permission.get("/v0/merged-person/self")
         assert response.status_code == 200
         result = response.json()
         assert result is not None
